Refactor daily pages with shared component
